api/src/api/requirements.txt
aiohttp~=3.9.2
pytest>=5.4.2
tortoise-orm~=0.21.6
redis~=5.0.8
quart~=0.19.6
quart-schema[pydantic]~=0.20.0